requirements.txt
urllib3[secure]
Django~=5.1
# M2Crypto # has been removed by commit "use hashlib+smime for upload verification mails instead of M2Crypto" (19.1.2018)
Markdown
# Warning: updating Pygments currently breaks the syntax highlighting for annotated solution files.
# DO NOT UPDATE YET!
Pygments~=2.7.0;python_version>="3.4"
chardet
django-extensions

# It appears that due to the pull-request https://github.com/aljosa/django-tinymce/pull/103 merged in django-tinymce==2.4.0, jquery is loaded after our onw jquery
# loaded in admin-sites. Using an old version of django-tinymce (2.3.x) is not possible because it uses methods removed between Django 1.8 and 1.11.
# We might want to use the django-admin provieded jquery in our admin-site jquery snippets?!?!?
django-tinymce~=4.1
django-debug-toolbar
docutils
psycopg2-binary
# selenium # has been removed by commit "remove selenium and phantomjs dependency" (16.1.2020)

#modules used in fix bug #255 in safeexec.py
psutil

#for LDAP support make sure, that ldap packages libsasl2-dev and libldap2-dev are installed on your operation system
python-ldap

#needed for praktomat.wsgi not only in unit tests for it
mod_wsgi # make sure, that you have called : sudo apt-get install apache2-dev

#needed for task statistic as boxplot diagrams
pandas
matplotlib
